+++ src/cmd/go/internal/imports/build.go
@@ -24,6 +24,7 @@ import (
 	"errors"
 	"fmt"
 	"go/build/constraint"
+	"internal/syslist"
 	"strings"
 	"unicode"
 )
@@ -213,7 +214,7 @@ func matchTag(name string, tags map[string]bool, prefer bool) bool {
 	case "darwin":
 		return tags["ios"]
 	case "unix":
-		return unixOS[cfg.BuildContext.GOOS]
+		return syslist.UnixOS[cfg.BuildContext.GOOS]
 	default:
 		return false
 	}
@@ -295,80 +296,14 @@ func MatchFile(name string, tags map[string]bool) bool {
 		l = l[:n-1]
 	}
 	n := len(l)
-	if n >= 2 && KnownOS[l[n-2]] && KnownArch[l[n-1]] {
+	if n >= 2 && syslist.KnownOS[l[n-2]] && syslist.KnownArch[l[n-1]] {
 		return matchTag(l[n-2], tags, true) && matchTag(l[n-1], tags, true)
 	}
-	if n >= 1 && KnownOS[l[n-1]] {
+	if n >= 1 && syslist.KnownOS[l[n-1]] {
 		return matchTag(l[n-1], tags, true)
 	}
-	if n >= 1 && KnownArch[l[n-1]] {
+	if n >= 1 && syslist.KnownArch[l[n-1]] {
 		return matchTag(l[n-1], tags, true)
 	}
 	return true
-}
-
-var KnownOS = map[string]bool{
-	"aix":       true,
-	"android":   true,
-	"darwin":    true,
-	"dragonfly": true,
-	"freebsd":   true,
-	"hurd":      true,
-	"illumos":   true,
-	"ios":       true,
-	"js":        true,
-	"linux":     true,
-	"nacl":      true, // legacy; don't remove
-	"netbsd":    true,
-	"openbsd":   true,
-	"plan9":     true,
-	"solaris":   true,
-	"wasip1":    true,
-	"windows":   true,
-	"zos":       true,
-}
-
-// unixOS is the set of GOOS values matched by the "unix" build tag.
-// This is not used for filename matching.
-// This is the same list as in go/build/syslist.go and cmd/dist/build.go.
-var unixOS = map[string]bool{
-	"aix":       true,
-	"android":   true,
-	"darwin":    true,
-	"dragonfly": true,
-	"freebsd":   true,
-	"hurd":      true,
-	"illumos":   true,
-	"ios":       true,
-	"linux":     true,
-	"netbsd":    true,
-	"openbsd":   true,
-	"solaris":   true,
-}
-
-var KnownArch = map[string]bool{
-	"386":         true,
-	"amd64":       true,
-	"amd64p32":    true, // legacy; don't remove
-	"arm":         true,
-	"armbe":       true,
-	"arm64":       true,
-	"arm64be":     true,
-	"ppc64":       true,
-	"ppc64le":     true,
-	"mips":        true,
-	"mipsle":      true,
-	"mips64":      true,
-	"mips64le":    true,
-	"mips64p32":   true,
-	"mips64p32le": true,
-	"loong64":     true,
-	"ppc":         true,
-	"riscv":       true,
-	"riscv64":     true,
-	"s390":        true,
-	"s390x":       true,
-	"sparc":       true,
-	"sparc64":     true,
-	"wasm":        true,
 }
